requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
pytz
rich
python-dotenv
requests